requirements.txt
beamngpy==1.24
click
colorama
cycler
descartes
fonttools
importlib-metadata
Jinja2
kiwisolver
MarkupSafe
matplotlib
msgpack
numpy
opencv-python
packaging
pandas
Pillow
PyOpenGL
pyparsing
python-dateutil
scipy
setuptools-scm
Shapely
tensorflow
tomli
zipp
deap>=1.3.3